requirements.txt
grpc
grpcio
gevent
thundersnow
futures
protobuf>=3.1.0
googleapis-common-protos>=1.5.0
requests
ipython
jinja2
faker
